Missed a dataremote intValue on the progress widget for devices. r=Mook on irc.
git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@9056 5aad4ff5-4b58-4159-ba3f-0e6328df4cad

+++ app/content/scripts/device/deviceProgress.js
@@ -265,7 +265,7 @@ var DPW = {
       this._dText1Remote.stringValue =
                 SBFormattedString("device.status.progress_complete_" +
                                     aOperationHead,
-                                  [this._totalItems]);
+                                  [this._totalItems.intValue]);
       this._dText2Remote.stringValue =
                                   SBString("device.status.progress_idle");
